remove F# code [prune]
I want to use the various facilities in Jetbrain Rider to help with
stripping down the code, but I can't get the F# code to compile and so
the IDE is cluttered with a bunch of errors. Rider just gives an error
about not being able to find/run fsc.exe, and no amount of messing
with my %PATH% in Windows seems to help.

The project builds fine when running build.cmd in a Windows command
prompt, so I know that the problem is specific to Rider.